Use LZMA compression to compress Protobuild executable
This update modifies Protobuild to compress and extract it's own code at runtime (moving most of the code into Protobuild.Internal).  In addition, it adds the LZMA and FastJSON libraries (the latter to be used with the new package API).

This results in a new Protobuild executable size of 128kb, which is smaller than the previous 171kb.

Because of this process, the build is now a 3 stage process; the first pass which compiles the GUI manager and recompresses build resources, the second pass which rebuilds Protobuild.Internal and then compresses it, and the final pass which builds the Protobuild.exe file itself (with the compressed embedded Protobuild.Internal.dll).

I've also cleaned up the build scripts in general, removing and consolidating the build scripts.  I've removed the build scripts for Windows for now, but this is only because I work on Linux and don't have a Windows machine available to write an equivalent Batch file for the new process.

+++ Protobuild.Tests/ProtobuildTest.cs
@@ -15,7 +15,7 @@
         protected void SetupTest(string name)
         {
             // This is used to ensure Protobuild.exe is referenced.
-            Console.WriteLine(typeof(ModuleInfo).FullName);
+            Console.WriteLine(typeof(Protobuild.Bootstrap.Program).FullName);
 
             this.m_TestName = name;
 
@@ -65,12 +65,20 @@
 
         protected void Generate(string platform = null, string args = null, bool expectFailure = false)
         {
-            this.PurgeSolutionsAndProjects(this.m_TestLocation);
+            this.OtherMode("generate", platform, args, expectFailure);
+        }
+
+        protected void OtherMode(string mode, string platform = null, string args = null, bool expectFailure = false, bool purge = true)
+        {
+            if (purge)
+            {
+                this.PurgeSolutionsAndProjects(this.m_TestLocation);
+            }
 
             var pi = new ProcessStartInfo
             {
                 FileName = Path.Combine(this.m_TestLocation, "Protobuild.exe"),
-                Arguments = "--generate " + (platform ?? "Windows") + " " + (args ?? string.Empty),
+                Arguments = "--" + mode + " " + (platform ?? "Windows") + " " + (args ?? string.Empty),
                 WorkingDirectory = this.m_TestLocation,
                 CreateNoWindow = true,
                 UseShellExecute = false,
